Protect Routes and add user_id support

+++ tasks-service/routes.py
@@ -16,9 +16,13 @@ class TaskList(Resource):
         current_user = get_jwt_identity()
         tasks = TaskModel.query.filter_by(user_id=current_user['id']).all() # Get Tasks of user
         return task_list_schema.dump(tasks)
+    
+    @jwt_required
     def post(self):
+        current_user = get_jwt_identity()
         new_task = TaskModel(
             title=request.json['title'],
+            user_id=current_user['id']
         )
         db.session.add(new_task)
         db.session.commit()
@@ -26,24 +30,40 @@ class TaskList(Resource):
 
 
 class TaskSingle(Resource):
+    @jwt_required
     def get(self, task_id):
+        current_user = get_jwt_identity()
         task = TaskModel.query.get_or_404(task_id)
-        return task_single_schema.dump(task), '200'
 
+        if task.user_id != current_user['id']:
+            return {'message': 'Unauthorized'}, '403'
+        else:
+            return task_single_schema.dump(task), '200'
+
+    @jwt_required
     def patch(self, task_id):
+        current_user = get_jwt_identity()
         task = TaskModel.query.get_or_404(task_id)
 
-        if 'title' in request.json:
-            task.title = request.json['title']
+        if task.user_id != current_user['id']:
+            return {'message': 'Unauthorized'}, '403'
+        else:
+            if 'title' in request.json:
+                task.title = request.json['title']
 
-        db.session.commit()
-        return task_single_schema.dump(task), '200'
+            db.session.commit()
+            return task_single_schema.dump(task), '200'
 
+    @jwt_required
     def delete(self, task_id):
+        current_user = get_jwt_identity()
         task = TaskModel.query.get_or_404(task_id)
-        db.session.delete(task)
-        db.session.commit()
-        return '', '204'
+        if task.user_id != current_user['id']:
+            return {'message': 'Unauthorized'}, '403'
+        else:
+            db.session.delete(task)
+            db.session.commit()
+            return '', '204'
 
 def init_api(app):
     api = Api(app)
